Working chat, and made something for the translationKey
