+++ cubic-server/CMakeLists.txt
@@ -79,3 +79,4 @@ add_subdirectory (generation)
 add_subdirectory (thread_pool)
 add_subdirectory (blocks)
 add_subdirectory (items)
+add_subdirectory (chat)
