+++ cubic-server/Player.hpp
@@ -106,6 +106,7 @@ private:
     void _onQueryBlockEntityTag(const std::shared_ptr<protocol::QueryBlockEntityTag> &pck);
     void _onChangeDifficulty(const std::shared_ptr<protocol::ChangeDifficulty> &pck);
     void _onChatMessage(const std::shared_ptr<protocol::ChatMessage> &pck);
+    void _onMessageAcknowledgement(const std::shared_ptr<protocol::MessageAcknowledgement> &pck);
     void _onChatCommand(const std::shared_ptr<protocol::ChatCommand> &pck);
     void _onClientCommand(const std::shared_ptr<protocol::ClientCommand> &pck);
     void _onClientInformation(const std::shared_ptr<protocol::ClientInformation> &pck);
