+++ cubic-server/chat/CMakeLists.txt
@@ -1,0 +1,8 @@
+target_sources (${CMAKE_PROJECT_NAME} PRIVATE
+    ClickEvent.hpp
+    ClickEvent.cpp
+    HoverEvent.hpp
+    HoverEvent.cpp
+    Message.cpp
+    Message.hpp
+)
