+++ cubic-server/protocol/serialization/pop.hpp
@@ -29,23 +29,6 @@ constexpr ClientInformationChatMode popClientInformationChatMode(uint8_t *&at, u
     return static_cast<ClientInformationChatMode>(value);
 }
 
-constexpr Instant popInstantJavaObject(uint8_t *&at, uint8_t *eof)
-{
-    Instant instant;
-    // Java class thingy, don't ask...
-    // See the java serialization spec if you like the pain
-    // https://docs.oracle.com/javase/8/docs/platform/serialization/spec/protocol.html
-    if (eof - at < 35)
-        throw PacketEOF("Not enough data in packet to parse an Instant");
-    at += 36;
-    uint8_t objectIdentifier = popByte(at, eof);
-    if (objectIdentifier != 0x02)
-        throw WrongObjectType("Wrong object identifier for Instant");
-    instant.seconds = popVarLong(at, eof);
-    instant.nanos = popVarInt(at, eof);
-    return instant;
-}
-
 constexpr ClientInformationMainHand popClientInformationMainHand(uint8_t *&at, uint8_t *eof)
 {
     auto value = popVarInt(at, eof);
